Internet-Engineering/crud_project_django/requirements.txt
asgiref==3.2.7
astroid==2.3.3
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
Django==3.0.5
django-crispy-forms==1.9.0
idna==2.9
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.3
pylint==2.4.4
pyparsing==2.4.7
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
wrapt==1.11.2
